Prepare and fit a text classification pipeline

Prepare and fit a text classification pipeline with Scikit-learn.

factory_pipeline_r(
  x,
  y,
  tknz = "spacy",
  ordinal = FALSE,
  metric = "class_balance_accuracy_score",
  cv = 5,
  n_iter = 2,
  n_jobs = 1,
  verbose = 3,
  learners = c("SGDClassifier", "RidgeClassifier", "Perceptron",
    "PassiveAggressiveClassifier", "BernoulliNB", "ComplementNB", "MultinomialNB",
    "RandomForestClassifier"),
  theme = NULL
)

Arguments

x
A data frame with the text feature.

y
A vector with the response variable.

tknz
Tokenizer to use ("spacy" or "wordnet").

ordinal
Whether to fit an ordinal classification model. The ordinal model is the implementation of Frank and Hall (2001) that can use any standard classification model that calculates probabilities.

metric
A string. Scorer to use during pipeline tuning ("accuracy_score", "balanced_accuracy_score", "matthews_corrcoef", "class_balance_accuracy_score").

cv
Number of cross-validation folds.

n_iter
Number of parameter settings that are sampled (see sklearn.model_selection.RandomizedSearchCV).

n_jobs
Number of jobs to run in parallel (see sklearn.model_selection.RandomizedSearchCV). NOTE: If your machine does not have the number of cores specified in n_jobs, then an error will be returned.

verbose
Controls the verbosity (see sklearn.model_selection.RandomizedSearchCV).

learners
A vector of Scikit-learn names of the learners to tune. Must be one or more of "SGDClassifier", "RidgeClassifier", "Perceptron", "PassiveAggressiveClassifier", "BernoulliNB", "ComplementNB", "MultinomialNB", "KNeighborsClassifier", "NearestCentroid", "RandomForestClassifier". When a single model is used, it can be passed as a string.

theme
For internal use by Nottinghamshire Healthcare NHS Foundation Trust or other trusts that use theme labels ("Access", "Environment/ facilities" etc.). The column name of the theme variable. Defaults to NULL. If supplied, the theme variable will be used as a predictor (along with the text predictor) in the model that is fitted with criticality as the response variable. The rationale is two-fold. First, to help the model improve predictions on criticality when the theme labels are readily available. Second, to force the criticality for "Couldn't be improved" to always be "3" in the training and test data, as well as in the predictions. This is the only criticality value that "Couldn't be improved" can take, so by forcing it to always be "3", we are improving model performance, but are also correcting possible erroneous assignments of values other than "3" that are attributed to human error.

Value

A fitted Scikit-learn pipeline containing a number of objects that can be accessed with the $ sign (see examples). For a partial list see "Atributes" in sklearn.model_selection.RandomizedSearchCV. Do not be surprised if more objects are in the pipeline than those in the aforementioned "Attributes" list. Python objects can contain a wide array of objects, from numeric results (e.g. the pipeline's accuracy), to methods (i.e. functions in the R lingo) and classes. In Python, these are normally accessed with object.<whatever>, but in R the command is object$<whatever>. For instance, one can access method predict() to make predictions on unseen data. See examples.

Details

The pipeline's parameter grid switches between two approaches to text classification: Bag-of-Words and Embeddings. For the former, both TF-IDF and raw counts are tried out.

The pipeline does the following:

Feature engineering:

Converts text into TF-IDFs or GloVe word vectors with spaCy.

Creates a new feature that is the length of the text in each record.

Performs sentiment analysis on the text feature and creates new features that are all scores/indicators produced by TextBlob and vaderSentiment.

Applies sklearn.preprocessing.KBinsDiscretizer to the text length and sentiment indicator features, and sklearn.preprocessing.StandardScaler to the embeddings (word vectors).

Up-sampling of rare classes: uses imblearn.over_sampling.RandomOverSampler to up-sample rare classes. Currently the threshold to consider a class as rare and the up-balancing values are fixed and cannot be user-defined.

Tokenization and lemmatization of the text feature: uses spaCy (default) or NLTK. It also strips punctuation, excess spaces, and metacharacters "r" and "n" from the text. It converts emojis into "__text__" (where "text" is the emoji name), and NA/NULL values into "__notext__" (the pipeline does get rid of records with no text, but this conversion at least deals with any escaping ones).

Feature selection: Uses sklearn.feature_selection.SelectPercentile with sklearn.feature_selection.chi2 for TF-IDFs or sklearn.feature_selection.f_classif for embeddings.

Fitting and benchmarking of user-supplied Scikit-learn estimators.

The numeric values in the grid are currently lists/tuples (Python objects) of values that are defined either empirically or are based on the published literature (e.g. for Random Forest, see Probst et al. 2019). Values may be replaced by appropriate distributions in a future release.

Note

The pipeline uses the tokenizers of Python library pxtextmining. Any warnings from Scikit-learn like UserWarning: The parameter 'token_pattern' will not be used since 'tokenizer' is not None' can therefore be safely ignored.

Also, any warnings about over-sampling can also be safely ignored. These warnings are simply a result of an internal check in the over-sampler of imblearn.

References

Frank E. & Hall M. (2001). A Simple Approach to Ordinal Classification. Machine Learning: ECML 2001 145--156.

Pedregosa F., Varoquaux G., Gramfort A., Michel V., Thirion B., Grisel O., Blondel M., Prettenhofer P., Weiss R., Dubourg V., Vanderplas J., Passos A., Cournapeau D., Brucher M., Perrot M. & Duchesnay E. (2011), Scikit-learn: Machine Learning in Python. Journal of Machine Learning Research 12:2825–-2830.

Probst P., Bischl B. & Boulesteix A-L (2018). Tunability: Importance of Hyperparameters of Machine Learning Algorithms. https://arxiv.org/abs/1802.09596

Examples

# Prepare training and test sets
data_splits <- pxtextmineR::factory_data_load_and_split_r(
  filename = pxtextmineR::text_data,
  target = "label",
  predictor = "feedback",
  test_size = 0.90) # Make a small training set for a faster run in this example

# Let's take a look at the returned list
str(data_splits)

#> List of 6
#>  $ x_train    :'data.frame':	929 obs. of  1 variable:
#>   ..$ predictor: chr [1:929] "Staff was nice." "A bit difficult to understand the staff when they talk to you wearing masks.  Have lost a few items since I been here as well." "There were odd things that I found disturbing" "Nothing" ...
#>   ..- attr(*, "pandas.index")=Int64Index([1116, 6904, 6642, 8361, 2488, 7989, 1716, 2237, 5699, 3859,
#>             ...
#>             3891, 7769, 2108, 4564, 4055,  682,  647, 4854, 3162, 7381],
#>            dtype='int64', length=929)
#>  $ x_test     :'data.frame':	8362 obs. of  1 variable:
#>   ..$ predictor: chr [1:8362] "Nothing all good" "Nothing" "No problems, felt looked after" "Rob, amazing mental health practitioner. Very kind, understand." ...
#>   ..- attr(*, "pandas.index")=Int64Index([8593, 6048, 5986,  859, 6784, 2277, 1378, 3834, 4280, 9275,
#>             ...
#>             2161, 4236, 4514, 2251, 5990, 2676, 1207, 2797, 3007, 2741],
#>            dtype='int64', length=8362)
#>  $ y_train    : Named chr [1:929] "Staff" "Communication" "Care received" "Couldn't be improved" ...
#>   ..- attr(*, "names")= chr [1:929] "1116" "6904" "6642" "8361" ...
#>  $ y_test     : Named chr [1:8362] "Couldn't be improved" "Couldn't be improved" "Miscellaneous" "Staff" ...
#>   ..- attr(*, "names")= chr [1:8362] "8593" "6048" "5986" "859" ...
#>  $ index_train: int [1:929] 1116 6904 6642 8361 2488 7989 1716 2237 5699 3859 ...
#>  $ index_test : int [1:8362] 8593 6048 5986 859 6784 2277 1378 3834 4280 9275 ...

# Fit the pipeline
pipe <- pxtextmineR::factory_pipeline_r(
  x = data_splits$x_train,
  y = data_splits$y_train,
  tknz = "spacy",
  ordinal = FALSE,
  metric = "accuracy_score",
  cv = 2, n_iter = 1, n_jobs = 1, verbose = 3,
  learners = "SGDClassifier"
)

# Mean cross-validated score of the best_estimator
pipe$best_score_

#> [1] 0.4154431

# Best parameters during tuning
pipe$best_params_

#> $sampling__kw_args
#> $sampling__kw_args$up_balancing_counts
#> [1] 300
#> 
#> 
#> $preprocessor__texttr__text__transformer__use_idf
#> [1] FALSE
#> 
#> $preprocessor__texttr__text__transformer__tokenizer
#> <pxtextmining.helpers.tokenization.LemmaTokenizer>
#> 
#> $preprocessor__texttr__text__transformer__preprocessor
#> <function text_preprocessor at 0x0000000063EB9F70>
#> 
#> $preprocessor__texttr__text__transformer__norm
#> NULL
#> 
#> $preprocessor__texttr__text__transformer__ngram_range
#> $preprocessor__texttr__text__transformer__ngram_range[[1]]
#> [1] 3
#> 
#> $preprocessor__texttr__text__transformer__ngram_range[[2]]
#> [1] 3
#> 
#> 
#> $preprocessor__texttr__text__transformer__min_df
#> [1] 1
#> 
#> $preprocessor__texttr__text__transformer__max_df
#> [1] 0.95
#> 
#> $preprocessor__texttr__text__transformer
#> TfidfVectorizer(max_df=0.95, ngram_range=(3, 3), norm=None,
#>                 preprocessor=<function text_preprocessor at 0x0000000063EB9F70>,
#>                 tokenizer=<pxtextmining.helpers.tokenization.LemmaTokenizer>,
#>                 use_idf=False)
#> 
#> $preprocessor__sentimenttr__scaler__scaler__n_bins
#> [1] 8
#> 
#> $preprocessor__sentimenttr__scaler__scaler
#> KBinsDiscretizer(n_bins=8, strategy='kmeans')
#> 
#> $preprocessor__lengthtr__scaler__scaler
#> KBinsDiscretizer(n_bins=3, strategy='kmeans')
#> 
#> $featsel__selector__score_func
#> <function chi2 at 0x0000000063E86C10>
#> 
#> $featsel__selector__percentile
#> [1] 100
#> 
#> $featsel__selector
#> SelectPercentile(percentile=100,
#>                  score_func=<function chi2 at 0x0000000063E86C10>)
#> 
#> $clf__estimator__penalty
#> [1] "elasticnet"
#> 
#> $clf__estimator__max_iter
#> [1] 10000
#> 
#> $clf__estimator__loss
#> [1] "log"
#> 
#> $clf__estimator__class_weight
#> NULL
#> 
#> $clf__estimator
#> SGDClassifier(loss='log', max_iter=10000, penalty='elasticnet')
#> 

# Make predictions
preds <- pipe$predict(data_splits$x_test)
head(preds)

#> [1] "Couldn't be improved" "Couldn't be improved" "Care received"       
#> [4] "Staff"                "Staff"                "Staff"               

# Performance on test set #
# Can be done using the pipe's score() method
pipe$score(data_splits$x_test, data_splits$y_test)

#> [1] 0.485171

# Or using dplyr
data_splits$y_test %>%
  data.frame() %>%
  dplyr::rename(true = '.') %>%
  dplyr::mutate(
    pred = preds,
    check = true == preds,
    check = sum(check) / nrow(.)
  ) %>%
  dplyr::pull(check) %>%
  unique()

#> [1] 0.485171