Split dataset into training and test sets

Splits the dataset with Scikit-learn and returns the train/test data and their row/position indices.

factory_data_load_and_split_r(
  filename,
  target,
  predictor,
  test_size = 0.33,
  reduce_criticality = FALSE,
  theme = NULL,
  python_setup = FALSE,
  sys_setenv = NULL,
  which_python = NULL,
  which_venv = NULL,
  venv_name = NULL
)

Arguments

filename
A data frame with the data (class and text columns), otherwise the dataset name (CSV), including full path to the data folder (if not in the project's working directory), and the data type suffix (".csv").

target
A string with the name of the response variable.

predictor
A string with the name of the predictor variable.

test_size
Numeric. Proportion of data that will form the test dataset.

reduce_criticality
Logical. For internal use by Nottinghamshire Healthcare NHS Foundation Trust or other trusts that hold data on criticality. If TRUE, then all records with a criticality of "-5" (respectively, "5") are assigned a criticality of "-4" (respectively, "4"). This is to avoid situations where the pipeline breaks due to a lack of sufficient data for "-5" and/or "5". Defaults to FALSE.

theme
A string. For internal use by Nottinghamshire Healthcare NHS Foundation Trust or other trusts that use theme labels ("Access", "Environment/ facilities" etc.). The column name of the theme variable. Defaults to NULL. If supplied, the theme variable will be used as a predictor (along with the text predictor) in the model that is fitted with criticality as the response variable. The rationale is two-fold. First, to help the model improve predictions on criticality when the theme labels are readily available. Second, to force the criticality for "Couldn't be improved" to always be "3" in the training and test data, as well as in the predictions. This is the only criticality value that "Couldn't be improved" can take, so by forcing it to always be "3", we are improving model performance, but are also correcting possible erroneous assignments of values other than "3" that are attributed to human error.

python_setup
A logical whether to set up the Python version, virtual environment etc. that can be controlled with arguments sys_setenv, which_python, which_venv and venv_name. These arguments will be ignored when python_setup is FALSE. The purpose of python_setup is that users may wish to control the Python parameters outside the actual function, for the session in general.

sys_setenv
A string in the form "path_to_python/python.exe", indicating which Python to use (e.g. from a virtual environment).

which_python
Same as sys_setenv.

which_venv
A string that can be "conda", "miniconda" or "python".

venv_name
String. The name of the virtual environment.

text_col_name
A string with the column name of the text variable.

Value

A list of length 6: x_train (data frame), x_test (data frame), y_train (character vector), y_test (character vector), index_train (integer vector), and index_test (integer vector). The row names (names) of x_train and x_test (y_train and y_test) are index_train and index_test respectively.

References

Pedregosa F., Varoquaux G., Gramfort A., Michel V., Thirion B., Grisel O., Blondel M., Prettenhofer P., Weiss R., Dubourg V., Vanderplas J., Passos A., Cournapeau D., Brucher M., Perrot M. & Duchesnay E. (2011), Scikit-learn: Machine Learning in Python. Journal of Machine Learning Research 12:2825--2830

Examples

# One can set the python.exe and virtual environment directly in the pxtextmineR
# functions or globally, with experienceAnalysis
# (https://github.com/CDU-data-science-team/experienceAnalysis).

experienceAnalysis::prep_python(
  sys_setenv = "C:/Users/andreas.soteriades/Anaconda3/envs/pxtextmining_venv/python.exe",
  which_python = "C:/Users/andreas.soteriades/Anaconda3/envs/pxtextmining_venv/python.exe",
  which_venv = "conda",
  venv_name = "pxtextmining_venv"
)

data_splits <- pxtextmineR::factory_data_load_and_split_r(
  filename = text_data,
  target = "label",
  predictor = "feedback",
  test_size = 0.33)

# Let's take a look at the returned list
str(data_splits)

#> List of 6
#>  $ x_train    :'data.frame':	6224 obs. of  1 variable:
#>   ..$ predictor: chr [1:6224] "Nothing, because you saved my life." "Everything was good" "Staff are really good & good to talk to" "Everything is good but * hand very well , I think she is the best." ...
#>   ..- attr(*, "pandas.index")=Int64Index([8820, 5390,  959, 5364, 3414, 8387, 2337, 8169, 6172, 4516,
#>             ...
#>             3055, 6415, 4616, 5447, 7853, 5189, 4042, 3697, 4419, 5497],
#>            dtype='int64', length=6224)
#>  $ x_test     :'data.frame':	3067 obs. of  1 variable:
#>   ..$ predictor: chr [1:3067] "Excellent" "Help keep a healthy mind & body." "Everything was good, very quick access, staff were very supportive, and put me at ease throughout." "They are very caring,kind,helpful and amazing team" ...
#>   ..- attr(*, "pandas.index")=Int64Index([5401, 1274, 1421, 6645, 2875, 1180, 5087, 5581,  274, 1166,
#>             ...
#>             8813, 7696, 2952, 1153, 5337, 5535, 5126, 9054, 7049,  754],
#>            dtype='int64', length=3067)
#>  $ y_train    : Named chr [1:6224] "Care received" "Miscellaneous" "Staff" "Miscellaneous" ...
#>   ..- attr(*, "names")= chr [1:6224] "8820" "5390" "959" "5364" ...
#>  $ y_test     : Named chr [1:3067] "Care received" "Care received" "Access" "Staff" ...
#>   ..- attr(*, "names")= chr [1:3067] "5401" "1274" "1421" "6645" ...
#>  $ index_train: int [1:6224] 8820 5390 959 5364 3414 8387 2337 8169 6172 4516 ...
#>  $ index_test : int [1:3067] 5401 1274 1421 6645 2875 1180 5087 5581 274 1166 ...

# Each record in the split data is tagged with the row index of the original dataset
head(rownames(data_splits$x_train))

#> [1] "8820" "5390" "959"  "5364" "3414" "8387"

head(names(data_splits$y_train))

#> [1] "8820" "5390" "959"  "5364" "3414" "8387"

# Note that, in Python, indices start from 0 and go up to number_of_records - 1
all_indices <- data_splits$y_train %>%
  names() %>%
  c(names(data_splits$y_test)) %>%
  as.numeric() %>%
  sort()
head(all_indices) # Starts from zero

#> [1] 0 1 2 3 4 5

tail(all_indices) # Ends in nrow(text_data) - 1

#> [1] 9285 9286 9287 9288 9289 9290

length(all_indices) == nrow(text_data)

#> [1] TRUE